+++ web/src/Components/CMAForm/DescriptionOfProjectActivity.tsx
@@ -14,7 +14,7 @@ import GetLocationMapComponent from '../Maps/GetLocationMapComponent';
 import moment from 'moment';
 import { getBase64 } from '../../Definitions/Definitions/programme.definitions';
 import { RcFile } from 'antd/lib/upload';
-import { PURPOSE_CREDIT_DEVELOPMENT } from '../SLCFProgramme/AddNewProgramme/SLCFProgrammeCreationComponent';
+import { PURPOSE_CREDIT_DEVELOPMENT } from '../AddNewProgramme/ProgrammeCreationComponent';
 import LabelWithTooltip, { TooltipPostion } from '../LabelWithTooltip/LabelWithTooltip';
 import { CMASectoralScope } from '../../Definitions/Enums/programmeStage.enum';
