+++ web/src/Components/MonitoringReport/MonitoringReport.scss
